\name{adj.cssls}
\alias{adj.cssls}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Solve matrix equation
}
\description{
This function was obtained from NMF package (Gaujoux R., BMC Bioinformatics 2010,11:367) at https://github.com/renozao/NMF/blob/master/R/algorithms-snmf.R and was adjusted for some minor bug fixes
}
\usage{
adj.cssls(CtC, CtA, Pset = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{CtC}{
CtC
}
  \item{CtA}{
CtA
}
  \item{Pset}{
Pset
}
}

\references{
The origibal matlab code was proposed by M. H. Van Benthem and M. R. Keenan, J. Chemometrics 2004; 18: 441-450
}
\author{
Prabhakar Chalise, Rama Raghavan, Brooke Fridley
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
