\name{logit}
\alias{logit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
logit
}
\description{
logit function
}
\usage{
logit(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x
}
}

\author{
Prabhakar Chalise, Rama Raghavan, Brooke Fridley
}

\examples{

## The function is currently defined as
function (x) 
log(x) - log(1 - x)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
