\name{W.fcnnls}
\alias{W.fcnnls}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to calculate common W using Nonnegative Least Square method
}
\description{
This function was obtained from NMF package (Gaujoux R., BMC Bioinformatics 2010,11:367) at https://github.com/renozao/NMF/blob/master/R/algorithms-snmf.R and was modified for using multiple data sets.
}
\usage{
W.fcnnls(x = x, y = y, verbose = FALSE, pseudo = FALSE, eps = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x
}
  \item{y}{
y
}
  \item{verbose}{
verbose
}
  \item{pseudo}{
pseudo
}
  \item{eps}{
eps
}
}

\author{
Prabhakar Chalise, Rama Raghavan, Brooke Fridley
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
